+++ internals/mockdata/games.js
@@ -9,39 +9,13 @@ module.exports = [
         processing: false,
         turn: 22,
         turnDuration: 10,
+        dimensions: 10,
         turnDue: moment()
             .add(10, "minutes")
             .toISOString(),
-        //maxPlayers: 20,
-        startDate: "2017-06-19T03:15:05.654Z"
-    },
-    {
-        _id: "598b652d0404e604046bd35c",
-        number: 2,
-        active: true,
-        canEnlist: false,
-        processing: false,
-        turn: 125,
-        turnDuration: 15,
-        turnDue: moment()
-            .add(15, "minutes")
-            .toISOString(),
-        //maxPlayers: 10,
-        startDate: "2017-05-02T11:15:05.654Z"
-    },
-    {
-        _id: "598b652d0404e604046bd35d",
-        number: 3,
-        active: true,
-        canEnlist: false,
-        processing: true,
-        turn: 9865,
-        turnDuration: 10,
-        turnDue: moment()
-            .add(10, "minutes")
-            .toISOString(),
-        //maxPlayers: 500,
-        startDate: "2017-07-29T21:15:05.654Z"
+        startDate: moment()
+            .subtract(1, "days")
+            .toISOString()
     },
     {
         _id: "598b652d0404e604046bd35e",
@@ -51,13 +25,12 @@ module.exports = [
         processing: false,
         turn: 0,
         turnDuration: 15,
-        maxPlayers: 4,
-        dimensions: 10,
+        dimensions: 20,
         startDate: moment()
-            .add(4, "minutes")
+            .add(2, "months")
             .toISOString(),
         turnDue: moment()
-            .add(4, "minutes")
+            .add(2, "months")
             .toISOString()
     }
 ];