+++ internals/mockdata/players.js
@@ -1,5 +1,14 @@
 module.exports = [
     {
+        // g1
+        _id: "5990c7871e8258089436d748",
+        game: "598b652d0404e604046bd35b",
+        user: "597c523b7331ca2e2436b37f",
+        name: "Testempire Part Two Reloaded",
+        ticker: "TEST",
+        created: "2017-08-14T19:31:10.861Z"
+    },
+    {
         // g4
         _id: "59904696a7dba10320d3a092",
         game: "598b652d0404e604046bd35e",
@@ -7,14 +16,5 @@ module.exports = [
         name: "Empire of Mankind",
         ticker: "EOM",
         created: "2017-08-13T12:31:10.861Z"
-    },
-    {
-        // g2
-        _id: "5990c7871e8258089436d748",
-        game: "598b652d0404e604046bd35c",
-        user: "597c523b7331ca2e2436b37f",
-        name: "Testempire Part Two Reloaded",
-        ticker: "TEST",
-        created: "2017-08-14T19:31:10.861Z"
     }
 ];